9.1-9.4 UC

+++ src/main/java/com/SAS/League/Budget.java
@@ -5,7 +5,7 @@ import com.SAS.team.Team;
 public class Budget {
     private Team team;
     private Season season;
-    private int budget;
+    private double budget;
 
     /**
      * GamesArrangement constructor with attributes
@@ -14,7 +14,7 @@ public class Budget {
      * @param season: the budget's season
      * @param budget: the list of the games
      */
-    public Budget(Team team, Season season, int budget) {
+    public Budget(Team team, Season season, double budget) {
         this.team = team;
         this.season = season;
         this.budget = budget;
@@ -40,7 +40,7 @@ public class Budget {
     /**
      * @return the size of the budget
      */
-    public int getBudget() {
+    public double getBudget() {
         return budget;
     }
